Move qwant provider to app
This provider is an "addon", move providers outside api to be lighter for the futur.
Remove unused var
