+++ app/providers/mongo/image.providers.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/smileinnovation/imannotate/api/project"
 	"github.com/smileinnovation/imannotate/api/providers"
+	"github.com/smileinnovation/imannotate/app/providers/qwant"
 )
 
 var imageProviders = make(map[string]providers.ImageProvider)
@@ -22,7 +23,7 @@ func createImageProvider(prj *project.Project) {
 	log.Println("Provider for project", prj.Name, prj.ImageProvider, prj.ImageProviderOptions)
 	switch prj.ImageProvider {
 	case "qwant":
-		provider := providers.NewQwant(prj.ImageProviderOptions["qwantQuery"])
+		provider := qwant.NewQwant(prj.ImageProviderOptions["qwantQuery"])
 		imageProviders[prj.Name] = provider
 	}
 }
